+++ frontend/src/components/InspectorPanel/settings/schedule/utils/getTimeZone.ts
@@ -1,6 +1,4 @@
-export function getTimeZoneAbbreviation(
-  date: Date = new Date(),
-): string | null {
+export function getTimeZoneAbbreviation(): string | null {
   try {
     const formatter = new Intl.DateTimeFormat("en-US", {
       timeZone: Intl.DateTimeFormat().resolvedOptions().timeZone,
@@ -8,7 +6,7 @@ export function getTimeZoneAbbreviation(
       hour: "2-digit",
       minute: "2-digit",
     });
-    const parts = formatter.formatToParts(date);
+    const parts = formatter.formatToParts(new Date());
     return parts.find((part) => part.type === "timeZoneName")?.value ?? null;
   } catch {
     return null;
